+++ app.ts
@@ -1,0 +1,19 @@
+require("dotenv").config();
+import express from "express";
+import cors from "cors";
+import todoRoutes from "./src/todo/todo.routes";
+import roleRoutes from "./src/role/role.routes";
+import userRoutes from "./src/user/user.routes";
+import CustomMiddleware from "./middleware";
+import DBInit from "./config/db-init";
+
+const app = express();
+app.use(cors());
+app.use(express.json());
+
+app.use("/todos", todoRoutes);
+app.use("/roles", roleRoutes);
+app.use("/users", userRoutes);
+
+CustomMiddleware(app);
+DBInit(app);
