+++ config/db-init.ts
@@ -1,7 +1,8 @@
-const sequelize = require("./sequelize.js");
-const { defineAssociations } = require("./associations.js");
+import sequelize from "./sequelize";
+import { defineAssociations } from "./associations";
+import { Express } from "express";
 
-function DBInit(app) {
+const DBInit = (app: Express) => {
   sequelize
     .authenticate()
     .then(() => {
@@ -17,6 +18,6 @@ function DBInit(app) {
     .catch((err) => {
       console.error("Unable to connect to the database:", err);
     });
-}
+};
 
-module.exports = { DBInit };
+export default DBInit;
