+++ src/todo/todo.routes.js
@@ -1,10 +1,10 @@
-const express = require("express");
-const router = express.Router();
-const todoController = require("./todo.controller.js");
+import express from "express";
+import { todoController } from "./todo.controller.js";
 
+const router = express.Router();
 router.get("/get", todoController.getTodos);
 router.post("/create", todoController.createTodo);
 router.patch("/update/:id", todoController.updateTodo);
 router.patch("/:id/archive", todoController.softDeleteTodo);
 
-module.exports = router;
+export default router;
